app/requirements.txt
bme280pi
kafka-python
